refactor(addressSearch): move searchResultSymbols to lib package
