+++ src/lib/searchResultSymbols.ts
